Switch to go modules; fix breaking changes
Some of the packages we were using have been pulled, so I had to fix a
call to the github API. TBD whether it breaks things. I _think_ what's
going to happen is that the --force flag is going to stop working, but
I'd need to dig into the api to know for sure.